+++ src/app/login-user/login-user.component.ts
@@ -9,6 +9,9 @@ import {
 
 import { ApiCallService } from '../Services/api-call.service';
 import { Router } from '@angular/router';
+import { DashboardService } from '../Services/dashboard.service';
+import { IsValidEmail, IsValidPassword } from '../Validation';
+import { ToastService } from '../Services/toast.service';
 
 @Component({
   selector: 'app-login-user',
@@ -22,18 +25,24 @@ export class LoginUserComponent implements OnInit {
   constructor(
     private formBuilder: FormBuilder,
     private apicall: ApiCallService,
-    private router : Router
+    private router: Router,
+    private share: DashboardService,
+    private toastService : ToastService
   ) {
     this.form = this.formBuilder.group({
-      User_Email: ['', [Validators.required, Validators.email]],
+      User_Email: ['', [Validators.required, Validators.pattern(IsValidEmail)]],
 
-      User_Password: ['', [Validators.required, Validators.minLength(6)]],
+      User_Password: [
+        '',
+        [Validators.required, Validators.pattern(IsValidPassword)],
+      ],
     });
   }
 
   ngOnInit(): void {
-    sessionStorage.clear();
-   
+    if (typeof window !== 'undefined') {
+      sessionStorage.clear();
+    }
   }
 
   get f(): { [key: string]: AbstractControl } {
@@ -41,7 +50,6 @@ export class LoginUserComponent implements OnInit {
   }
 
   onSubmit(): void {
-
     this.submitted = true;
 
     if (this.form.invalid) {
@@ -57,26 +65,23 @@ export class LoginUserComponent implements OnInit {
     };
     //user login
     this.apicall.userapiservice(JSON.stringify(data)).subscribe((result) => {
-  
       if (result != null && result != '' && result != undefined) {
         if (result['ID'] == '200') {
+          this.share.JWTToken = result['Message'];
           sessionStorage.clear();
-          sessionStorage.setItem("Role","User")
-          sessionStorage.setItem("IsLoggedIn",String(true))
+          sessionStorage.setItem('Role', 'User');
+          sessionStorage.setItem('Token', result['Message']);
+          sessionStorage.setItem('IsLoggedIn', String(true));
           this.router.navigate(['/dashboard-user']);
         } else {
           document.getElementById('failure').style.display = 'block';
           this.form.reset();
-          setTimeout(() => {
-            document.getElementById('failure').style.display = 'none';
-          }, 3000);
-       
+          this.toastService.show('Login failed', { classname: 'bg-danger text-light', delay: 2000 });
+
         }
       }
     });
 
     this.submitted = false;
   }
-
-
 }
